New Api Intrigation

+++ lib/data/models/product.dart
@@ -11,7 +11,7 @@ class Product {
   String? discountPrice;
   String? image;
   int? stock;
-  double? star;
+  int? star;
   String? remark;
   int? categoryId;
   int? brandId;
